// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  index: typeof index;
  insights: typeof insights;
  recordings: typeof recordings;
  streaming: typeof streaming;
  workflows: typeof workflows;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;